pyproject.toml
[project]
name = "pingpong-esp32"
version = "0.1.0"
requires-python = ">=3.10"
description = "Pingpong detector"
readme = "README.md"
authors = [{name="Zach Moshe"}]
dependencies = [
    "aiohttp>=3.13.0",
    "esptool>=5.1.0",
    "fastapi>=0.118.3",
    "ipykernel>=7.0.1",
    "ipywidgets>=8.1.7",
    "matplotlib>=3.10.7",
    "moviepy>=2.2.1",
    "mpremote>=1.26.1",
    "ngrok>=1.4.0",
    "numpy>=2.2.6",
    "pyserial>=3.5",
    "python-dotenv>=1.1.1",
    "scipy>=1.15.3",
    "slack-sdk>=3.37.0",
    "uvicorn>=0.37.0",
]

[project.optional-dependencies]
dev = [
  "mpremote>=1.23",
  "esptool>=4.7",
  "pyserial>=3.5",
  "rich>=13.7"
]

[tool.uv]
